% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbcon.R
\name{dbcon}
\alias{dbcon}
\alias{closeCon}
\alias{closeCon,MariaDBConnection-method}
\title{Connect to a database}
\usage{
dbcon(server = c("scidb", "scidb_replica"), db, driver = "MariaDB")

closeCon(con)

\S4method{closeCon}{MariaDBConnection}(con)
}
\arguments{
\item{server}{Character vector of server (group) names in the `my.cnf` file.
Servers are tried in order until a connection succeeds.}

\item{db}{Optional database name to `USE` after connecting.}

\item{driver}{Driver identifier. Currently supports `"MariaDB"` (default).}

\item{con}{The connection object.}

\item{...}{Further arguments forwarded to [DBI::dbConnect()].}
}
\value{
A DBI connection object.
}
\description{
`dbcon()` creates a DBI connection. By default it tries servers in order and
uses credentials and defaults from the `dbo` package options.
}
\details{
Connection parameters are taken from:
- `options(dbo.my.cnf)` as `default.file` (for login groups)
- `options(dbo.tz)` as `timezone`
}
\note{
`timezone` and `default.file` are set via `options()` (typically at package
startup). The selected `server` entry is passed as `group`.
}
\seealso{
[DBI::dbConnect()], [DBI::dbDisconnect()]
}
