% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{package}
\name{dbo}
\alias{dbo-package}
\alias{dbo}
\title{\strong{d}ata \strong{b}ase interface for scidb.\strong{o}rn.mpg.de}
\description{
An interface to \href{http://scidb.mpio.orn.mpg.de}{scidb.mpio.orn.mpg.de} databases
}
