% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbWrite.R
\name{dbUpdate}
\alias{dbUpdate}
\title{Update table}
\usage{
dbUpdate(con, dbtable, dat, by, which)
}
\arguments{
\item{con}{a connection object}

\item{dbtable}{dbtable name}

\item{dat}{a data.frame}

\item{by}{by which colum to update}

\item{which}{which colum to update}
}
\description{
Update database table using a data.frame. Only one column can be updated.
}
\examples{
\dontrun{

x = data.table(y = letters[1:5], pk = 1:5)
con = dbcon(db = 'tests')
dbWriteTable(con, "X", x, row.names = FALSE, overwrite =TRUE)
x1 = x[1:2][,y := c("Z", "W")]
o = dbUpdate(con, "X", x1, by = "pk", which = "y")
dbq(con, "SELECT * FROM X")
dbExecute(con, "DROP table X")
closeCon(con)

}

}
