% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbGet.R
\name{get_table_comments}
\alias{get_table_comments}
\title{Table comments}
\usage{
get_table_comments(con, tableName)
}
\arguments{
\item{con}{a *Connection object. If missing dbq() without a connection argument is called.}

\item{tableName}{table name}
}
\description{
Extract table comments to data.table
}
\examples{
\dontrun{
get_table_comments(con, tableName = '')
con = dbcon(host = '127.0.0.1', db = 'tests')
dbq(con, 'CREATE TABLE temp (f1 VARCHAR(50) NULL,f2 FLOAT NULL)' )
dbInsertInto(con, 'temp', x)
dbq(con, 'DROP TABLE temp')
}
}
