% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgroups_AVES.R
\name{symerge}
\alias{symerge}
\title{Synonyms merge}
\usage{
symerge(ttax, tdata, sy, clean = TRUE)
}
\arguments{
\item{ttax}{table containing the taxonomy we wish to keep.}

\item{tdata}{table containing the data to add to ttax.}

\item{sy}{the synonyms table. A table with two columns (scinam, syid), see examples.}

\item{clean}{default to TRUE; remove scinam_tdata, and renames scinam_ttax to scinam}
}
\description{
Synonyms merge
}
\examples{
ttax  = data.table(scinam = c('a', 'b', 'c', 'd', 'm', 'n'), V1 = rnorm(6))
tdata  = data.table(scinam = c('a', 'b', 'j', 'k', 'n', 'm'), V2 = rnorm(6))
sy = data.table(scinam = c('a', 'b', 'c', 'j', 'd', 'k', 'm', 'n'), syid = c(1,2,3,3,4,4,5,5))
symerge(ttax, tdata, sy)

\dontrun{
require(dbo)   
tdata = dbq( q = 'select scinam, bownam from AVES_taxonomy.bow')
ttax  = dbq( q = 'select scinam, family from AVES_taxonomy.birdtree')
sy    = dbq( q= 'select scinam, syid from  AVES_taxonomy.synonyms_v3')

x = symerge(ttax, tdata, sy)

x = symerge(ttax, tdata, sy, clean = FALSE)

}
}
