% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbGet.R
\name{tableComments}
\alias{tableComments}
\title{Get table comments}
\usage{
tableComments(con, tableName, db)
}
\arguments{
\item{con}{a *Connection object. See \code{\link[=dbcon]{dbcon()}}}

\item{tableName}{table name}

\item{db}{optional database name.}
}
\description{
Extract table comments to data.table
}
