% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_db}
\alias{test_db}
\title{Test database}
\usage{
test_db(user = "testuser", host = "127.0.0.1", db = "tests", pwd)
}
\arguments{
\item{user}{default to 'testuser'}

\item{host}{default to '127.0.0.1'}

\item{db}{default to 'tests'}

\item{pwd}{pwd}
}
\description{
Installs a mariadb test database with spatial support.
}
\note{
test_db() works independently on other functions in dbo
}
