% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbcon.R
\name{dbcon}
\alias{dbcon}
\alias{closeCon}
\alias{closeCon,MariaDBConnection-method}
\title{Connect to a database}
\usage{
dbcon(driver = "MariaDB", server = "scidb", db, config.file, ...)

closeCon(con)

\S4method{closeCon}{MariaDBConnection}(con)
}
\arguments{
\item{driver}{MariaDB or mysql_gdal, ... Defaults to MariaDB}

\item{db}{active database name}

\item{...}{pass to dbConnect}
}
\value{
a connection object
}
\description{
\code{dbcon} returns a database connection.
                  If user and password are not given dbq looks for TODO
}
\seealso{
\code{\link{xxxxxxx}}
}
