% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_dbcon.R
\name{dbcon}
\alias{dbcon}
\alias{closeCon}
\alias{closeCon,MariaDBConnection-method}
\title{Connect to a database}
\usage{
dbcon(driver = "MariaDB", server = "scidb", db, config.file, ...)

closeCon(con)

\S4method{closeCon}{MariaDBConnection}(con)
}
\arguments{
\item{driver}{MariaDB, defaults to MariaDB}

\item{server}{server name.}

\item{db}{active database name}

\item{config.file}{path to a config file. See \code{\link[=my.cnf]{my.cnf()}}}

\item{...}{further arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{con}{The connection object.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a connection object
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\code{dbcon} returns a database connection.
              If user and password are not given dbq looks for TODO
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    [DBI::dbConnect()]
}\if{html}{\out{</div>}}
}
