% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio_credentials.R
\name{my.cnf}
\alias{my.cnf}
\title{manage database credentials
manage database credentials for for different hosts .}
\usage{
my.cnf(delete = FALSE)
}
\arguments{
\item{delete}{removes the config file (default to FALSE).}
}
\description{
manage database credentials
manage database credentials for for different hosts .
}
\note{
If you are also using my.cnf with cli tools that use -`-defaults-group-suffix=`, make sure that any `[name]` entry is duplicated as `[clientname]`.
}
